requirements.txt
pyzmq
pyyaml
appdirs
attrdict
circus>=0.12.0
